requirements.txt
django==4.0
djangorestframework==3.12
psycopg2==2.9.5
drf-yasg
PyJWT==2.6.0
djangorestframework-simplejwt